+++ src/atoms/RadioButton/RadioButton.web.js
@@ -1,3 +1,5 @@
-const RadioButton = () => null;
+import React from 'react';
+
+const RadioButton = () => <div>WIP</div>;
 
 export default RadioButton;
